Refactored project
