+++ src/utils/project_types.js
@@ -1,0 +1,3 @@
+export const WEB_APP = "website";
+export const MOBILE_APP = "mobile";
+export const DESKTOP_APP = "desktop";
